requirements-dev.txt
ansible-core>=2.16.0
ansible-lint>=6.21.0
antsibull-docs
docker>=7
hvac
jmespath
yamllint==1.35.1
pre-commit
requests<2.32.0
-r requirements-test.txt
